A few fixes to previous ResUse de-MySQLization.
git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@1967 5946f024-35f4-0310-b1b3-9a3be3380cfb

+++ lib/dbdriver.h
@@ -213,7 +213,7 @@ public:
 	/// \brief Jumps to the given field within the result set
 	///
 	/// Wraps \c mysql_field_seek() in MySQL C API.
-	void field_seek(MYSQL_RES* res, int field) const
+	void field_seek(MYSQL_RES* res, size_t field) const
 			{ mysql_field_seek(res, field); }
 
 	/// \brief Return the connection options object
